refactor: co-locate unit tests with source files
Move tests/core/*.test.ts to src/core/*.test.ts.
Update vitest include to src/**/*.test.ts + tests/**/*.test.ts.
Exclude *.test.ts from tsconfig (partial Model objects).
Add mockReset: true to vitest config.
