+++ .changeset/colocate-tests.md
@@ -1,0 +1,8 @@
+---
+"@aliou/pi-ts-aperture": patch
+---
+
+Co-locate unit tests with source files.
+
+Moves core unit tests from tests/core/ to src/core/*.test.ts.
+
